+++ examples/futures/FuturesMarketData/GetMarkPriceKline.php
@@ -9,7 +9,14 @@ require_once __DIR__ . '/../../../vendor/autoload.php';
 $APIContract = new APIContractMarket(new CloudConfig([
     'url' => CloudConst::API_URL_V2_PRO,
     'timeoutSecond' => 5,
-]));;
+    'logger' => [
+        'enabled' => true,
+        'format' => 'default',
+        'outputToConsole' => true,
+    ] 
+]));
+
+
 $endTime = round(microtime(true));
 $startTime = $endTime - (60*60);
 $response = $APIContract->getContractMarkPriceKline(