+++ examples/futures/FuturesMarketData/GetMarketDepth.php
@@ -9,7 +9,12 @@ require_once __DIR__ . '/../../../vendor/autoload.php';
 $APIContract = new APIContractMarket(new CloudConfig([
     'url' => CloudConst::API_URL_V2_PRO,
     'timeoutSecond' => 5,
-]));;
+    'logger' => [
+        'enabled' => true,
+        'format' => 'default',
+        'outputToConsole' => true,
+    ] 
+]));
 
 $response = $APIContract->getContractDepth("BTCUSDT")['response'];