+++ examples/futures/Trading/GetPositionMode.php
@@ -1,10 +1,13 @@
 <?php
 
+use BitMart\CloudConst;
 use BitMart\Futures\APIContractTrading;
+use BitMart\Lib\CloudConfig;
 require_once __DIR__ . '/../../ExampleConfig.php';
 require_once __DIR__ . '/../../../vendor/autoload.php';
 
 $APIContract = new APIContractTrading(ExampleConfig::getExampleConfig());
 
-$response = $APIContract->getContractAssets()['response'];
+$response = $APIContract->getPositionMode()['response'];
 echo json_encode($response);
+