+++ examples/futures/Trading/GetPositionRisk.php
@@ -1,17 +1,11 @@
 <?php
 
-use BitMart\CloudConst;
 use BitMart\Futures\APIContractTrading;
-use BitMart\Lib\CloudConfig;
 
+require_once __DIR__ . '/../../ExampleConfig.php';
 require_once __DIR__ . '/../../../vendor/autoload.php';
 
-$APIContract = new APIContractTrading(new CloudConfig([
-    'url' => CloudConst::API_URL_V2_PRO,
-    'accessKey' => "<your_api_key>",
-    'secretKey' => "<your_secret_key>",
-    'memo' => "<your_memo>",
-]));
+$APIContract = new APIContractTrading(ExampleConfig::getExampleConfig());
 
 $response = $APIContract->getContractPositionRisk(
 )['response'];
@@ -21,6 +15,7 @@ echo json_encode($response);
 $response = $APIContract->getContractPositionRisk(
     [
         'symbol' => 'BTCUSDT',
+        'account' => 'futures',
     ]
 )['response'];
 echo json_encode($response);