+++ examples/spot/Trading/NewOrder.php
@@ -1,15 +1,12 @@
 <?php
-use BitMart\Lib\CloudConfig;
 use BitMart\Spot\APISpot;
 
 require_once __DIR__ . '/../../../vendor/autoload.php';
+require_once __DIR__ . '/../../ExampleConfig.php';
 
-$APISpot = new APISpot(new CloudConfig([
-    'accessKey' => "<your_api_key>",
-    'secretKey' => "<your_secret_key>",
-    'memo' => "<your_memo>",
-]));
+$APISpot = new APISpot(ExampleConfig::getExampleConfig());
 
+// Submit limit order with stpMode
 $response = $APISpot->postSubmitOrder(
     'BTC_USDT',
     'buy',
@@ -17,23 +14,23 @@ $response = $APISpot->postSubmitOrder(
     [
         'size' => '0.1',
         'price' => '8800',
-        'client_order_id' => 'test20000000005'
+        'client_order_id' => 'test20000000005',
+        'stpMode' => "cancel_maker"
     ]
 )['response'];
 
-echo json_encode($response);
+echo json_encode($response) . PHP_EOL;
 
-
+// Submit market order
 $response = $APISpot->postSubmitOrder(
     'BTC_USDT',
     'buy',
     'market',
     [
         'notional' => '2',
-        'price' => '8800',
         'client_order_id' => 'test20000000006'
     ]
 )['response'];
 
-echo json_encode($response);
+echo json_encode($response) . PHP_EOL;