+++ src/BitMart/Lib/CloudConfig.php
@@ -5,6 +5,7 @@ namespace BitMart\Lib;
 
 
 use BitMart\CloudConst;
+use BitMart\Lib\CloudLogger;
 
 class CloudConfig
 {
@@ -13,8 +14,9 @@ class CloudConfig
     public $secretKey;
     public $memo;
     public $timeoutSecond;
-    public $xdebug;
     public $customHeaders;
+    public $loggerConfig;
+    public $logger; // CloudLogger instance
 
     public function __construct($args = array())
     {
@@ -23,8 +25,11 @@ class CloudConfig
         $this->secretKey   = $args['secretKey'] ?? "your_api_secret_key";
         $this->memo        = $args['memo'] ?? "your_api_memo";
         $this->timeoutSecond          = $args['timeoutSecond'] ?? 5;
-        $this->xdebug      = $args['xdebug'] ?? false;
         $this->customHeaders      = $args['customHeaders'] ?? array();
+        $this->loggerConfig = $args['logger'] ?? [];
+        
+        // Create logger instance for this CloudConfig
+        $this->logger = new CloudLogger($this->loggerConfig);
     }